// frontend/src/generated/api.ts
export interface components {
    schemas: {
// ...
        FilterRuleResponse: {
            /** Format: date-time */
            created_at: string;
            is_positive: boolean;
            regex_pattern: string;
            /** Format: int32 */
            rule_id: number;
            /** Format: int32 */
            rule_order: number;
            /** Format: int32 */
            target_id?: number | null;
            target_name?: string | null;
            target_type: string;
            /** Format: date-time */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}